==> notebooks/02_data_exploration_Muhinyia.ipynb <==
# Data Narration and Metadata

#### Resources:

* [About Data](https://simulacrum.healthdatainsight.org.uk/wp-content/uploads/2023/04/Simulacrum-v2-User-Guide.pdf)
* [Data Resource Profile-NDRS](https://academic.oup.com/ije/article/49/1/16/5476570)
* [Data Resource Profile - SACT](https://academic.oup.com/ije/article-pdf/49/1/15/33650519/dyz137.pdf)

  
#### Acronyms and terms:

* NDRS - National Disease Registration Service.

* SACT — Systemic anti-cancer therapy treatments.

* CAS - Cancer Analysis System.

* RTDS - Radiotherapy Data Set -> Standard for radiotherapy data collection and submission.

* Regimen — a regimen is a single drug or a combination of different drugs given over a planned schedule for systemic anti-cancer therapy treatments and can include chemotherapy and hormone treatments.

* Episode -  a continuous period of care for radiotherapy including all preparation, planning and delivery of radiotherapy.

* Brachytherapy - s a procedure used to treat certain types of cancer and other conditions. It involves placing radioactive material inside the body. This is sometimes called internal radiation.

* Dosimetry - Amount of radiation energy deposited in a tissue.


## Data Structure
The Simulacrum v2 is made up of synthetic versions of the following datasets, held on the CAS:
#### 1. National Cancer Registration Dataset (NCRD)
-  Information about all patients diagnosed with a cancer in years 2016-2019 and their
tumour diagnoses
-  Patient and tumour tables

#### 2. Systemic Anti-Cancer Therapy (SACT) datasets
-  Information about all systemic anti-cancer therapy treatments received by patients
-  Regimen, cycle, drug detail and outcome tables

#### 3. Radiotherapy Dataset (RTDS)
- Information about all radiotherapy treatments received by patients
- Prescription and exposure tables

#### 4. Somatic genetic testing data
-  Information about somatic tests performed on patients’ tumours


## High-level description of data files

1. **sim_av_patient** — Patient-level info, e.g., ethnicity, gender, etc.; each row is a patient—identifier _PATIENTID_.

2. **sim_av_tumor** — Tumor-level info, e.g., cancer site, stage, and date of diagnosis, etc. (only for PRIMARY cancers); each row is a cancer tumor—identifier _TUMOURID_. (A patient can have more than one tumor record). 

3. **sim_sact_regimen** - Each SACT patient is assigned to a predetermined treatment plan that comprises at least one regimen. A _PATIENTID_ would show which patient is undertaking what regimen.

4. **sim_sact_cycle** - A regimen will be given across multiple cycles. This table reflects the frequency with which the course of treatment is repeated and the rest periods between administrations. A cycle links to a regimen through _MERGED_REGIMEN_ID_.

5. **sim_sact_drug_detail** - Records details about each drug administered within a cycle, e.g., dosage and method of delivery. A detail linked to a cycle through _MERGED_CYCLE_ID_.

6. **sim_sact_outcome** - Records the outcome of a particular regimen that was undertaken (which was completed or changed). The outcome could be, e.g., modified or stopped early. An outcome linked to a regimen via _MERGED_REGIMEN_ID_.


<u><b>Note</b></u>

A patient will often have one regimen per line of treatment, with multiple cycles recorded for the regimen and multiple drug administrations per cycle. However, this may not be the case. For example, it is known that:
- (i) some oral drugs are incorrectly prescribed as a single cycle for the duration of treatment.
- (ii) cycles for some regimens can become split and submitted as multiple different regimens.

This latter issue is understood to be more pronounced in earlier years prior to trusts adopting robust e-prescribing systems.

7. **sim_rtds_episode** - Each row is a separate episode, which is delivered over one or more appointments, i.e., each row is a separate episode-appointment. An episode will involve a single prescription of radiotherapy or more than one prescription sometimes. Episodes are linked to the patients through _PATIENTID_.

8. **sim_rtds_presicription** — A radiotherapy prescription relates to a series of identical doses of treatments to a single anatomical site. One prescription will be delivered over the course of one or more appointments, with one data row per unique prescription-appointment combination. Prescriptions can be for either teletherapy or Brachytherapy treatment. This table includes information on the anatomical site that received radiotherapy treatment if a patient has metastatic disease, the radiotherapy dose, and the treatment region.

9. **sim_rtds_exposure** - At each appointment for a prescription, patients receive a number of treatment exposures, where each exposure is the delivery of one radiotherapy beam. Each row is an exposure. An exposure could be a non-treatment one where a machine is switched on for treatment verification, or it could be a treatment one, i.e., dosimetry. Exposure relates to the patient via _PATIENT_ID_. Exposure has a one-to-many relationship with prescriptions via _PRESCRIPTIONID_, _RADIOTHERAPYEPISODEID_, _ATTENDID_, _LINKCODE_ and _PATIENTID_.

10. **sim_rtds_combined** - A combination of **sim_rtds_episode**, **sim_rtds_prescription** and **sim_rtds_exposure**. The table has one data row per exposure and contains all variables from across all the RTDS tables.

11. **sim_av_gene** - Contains information about molecular tests for somatic (acquired) changes in a tumour’s genetic material at the gene level. Each row represents a unique combination of TUMOURID and a tested gene/chromosome. Although multiple tests may be performed on a tumour gene, each gene will appear only once per tumor. Includes information on the type and date of test and test results for each gene tested. Where multiple tests have occurred, a hierarchy has been applied to report the test with the earliest
most definitive result. Genes arw linked to tumours in **sim_av_tumour** by _TUMOURID_.
    




 

    


# Potential Analysis questions

1. **Disparities at Diagnosis** (are certain ethnic groups at a higher risk?)

   - How does ethnicity correlate with key prognostic factors at the point of diagnosis? Are there significant differences in **age, PSA level, Gleason score, cancer stage** at diagnosis between major ethnic groups (e.g., White, Black, Asian)?
     

  
2.  **Differences in Treatment Outcomes**

    - Is there a persistent difference in x-year prostate-cancer-specific survival between different ethnic groups?
    (Does the treatment work eqaully for all/or does NHS need to give specialized attention to certain groups?)


3. **Gaps in Early Detection** - (As NHS plan mentioned priotising prevention over treatment)

   - How do socioeconomic status (as measured by the Index of Multiple Deprivation) and ethnicity correlate with key prognostic indicators at diagnosis, specifically Gleason score and clinical (TNM) stage?

# Data Fields


PATIENTID - Pseudonymised patient ID

GENDER - Person stated gender

ETHNICITY - Ethnicity

DEATHCAUSECODE_1A - As provided with death notification

DEATHCAUSECODE_1B - As provided with death notification

DEATHCAUSECODE_1C - As provided with death notification

DEATHCAUSECODE_2 - As provided with death notification

DEATHCAUSECODE_UNDERLYING - As provided with death notification

DEATHLOCATIONCODE - Code of the location (type) where the patient died, e.g. patients home, hospice etc.

VITALSTATUS - Vital status of the patient

VITALSTATUSDATE - Date of vital status

LINKNUMBER - Link number, substitute for NHS number in the real data, which is coded as NHSNUMBER.


PATIENTID - Pseudonymised patient ID

PRESCRIPTIONID - The identifier of a prescription

RADIOTHERAPYEPISODEID - The identifier of an episode

ATTENDID - The identifier of an attendance

APPTDATE - The date of an attendance

LINKCODE - Link code

RTTREATMENTMODALITY - The radiotherapy treatment modality, i.e., the type of treatment delivered during a radiotherapy prescription

RTPRESCRIBEDDOSE - The total prescribed absorbed radiation dose

PRESCRIBEDFRACTIONS - The prescribed number of radiotherapy fractions delivered to a patient in the radiotherapy prescription

RTACTUALDOSE - The total actual absorbed radiation dose

RTACTUALFRACTIONS - The delivered number of radiotherapy fractions delivered to a patient in the radiotherapy prescription

TREATMENTREGION - The specific region of the body being treated with radiotherapy.

TREATMENTANATOMICALSITE - A classification of part of the body to which the radiation dose is administered.

DECISIONTOTREATDATE - The date on which it was decided that the patient required a specific planned cancer treatment. This is the date that the consultation between the patient and the clinician took place and a planned cancer treatment was agreed.

EARLIESTCLINAPPROPDATE - The earliest date that it was clinically appropriate for an activity to take place. This is the same as the DECISIONTOTREATDATE unless there was an elective delay.

RADIOTHERAPYPRIORITY - The priority for the radiotherapy treatment course as classified by the requesting clinician

RADIOTHERAPYINTENT - The intent of the radiotherapy treatment

RADIOISOTOPE - The type of radioactive source used to deliver radiotherapy

RADIOTHERAPYBEAMTYPE - The prescribed type of beam for a radiotherapy exposure

RADIOTHERAPYBEAMENERGY - The prescribed energy of a radiotherapy exposure

TIMEOFEXPOSURE - The time when the radiotherapy exposure was first initiated on this attendance.


PRESCRIPTIONID - The identifier of a prescription

RADIOISOTOPE - The type of radioactive source used to deliver radiotherapy

RADIOTHERAPYBEAMTYPE - The prescribed type of beam for a radiotherapy exposure

RADIOTHERAPYBEAMENERGY - The prescribed energy of a radiotherapy exposure

TIMEOFEXPOSURE - The time when the radiotherapy exposure was first initiated on this attendance.

RADIOTHERAPYEPISODEID - The identifier of an episode

ATTENDID - The identifier of an attendance

APPTDATE - The date of an attendance


RTTREATMENTMODALITY - The radiotherapy treatment modality, i.e., the type of treatment delivered during a radiotherapy prescription

RTPRESCRIBEDDOSE - The total prescribed absorbed radiation dose

PRESCRIBEDFRACTIONS - The prescribed number of radiotherapy fractions delivered to a patient in the radiotherapy prescription

RTACTUALDOSE - The total actual absorbed radiation dose

RTACTUALFRACTIONS - The delivered number of radiotherapy fractions delivered to a patient in the radiotherapy prescription

TREATMENTREGION - The specific region of the body being treated with radiotherapy.

TREATMENTANATOMICALSITE - A classification of part of the body to which the actual dose is administered.

RADIOTHERAPYEPISODEID - The identifier of an episode


RADIOTHERAPYEPISODEID - The identifier of an episode

ATTENDID - The identifier of an attendance

APPTDATE - The date of an attendance

LINKCODE - Link code

DECISIONTOTREATDATE - The date on which it was decided that the patient required a specific planned cancer treatment. This is the date that the consultation between the patient and the clinician took place and a planned cancer treatment was agreed.

EARLIESTCLINAPPROPDATE - The earliest date that it was clinically appropriate for an activity to take place. This is the same as the DECISIONTOTREATDATE unless there was an elective delay.

RADIOTHERAPYPRIORITY - The priority for the radiotherapy treatment course as classified by the requesting clinician

RADIOTHERAPYINTENT - The intent of the radiotherapy treatment


MERGED_REGIMEN_ID - Pseudonymised regimen ID

DATE_OF_FINAL_TREATMENT - Start date (final therapy)

REGIMEN_MOD_DOSE_REDUCTION - Regimen modification indicator (dose reduction)

REGIMEN_MOD_TIME_DELAY - Regimen modification indicator (time delay)

REGIMEN_MOD_STOPPED_EARLY - Regimen modification indicator (days reduced)

REGIMEN_OUTCOME_SUMMARY - Planned treatment change reason


MERGED_DRUG_DETAIL_ID - Pseudonymised drug administration ID

MERGED_CYCLE_ID - Pseudonymised cycle ID

ACTUAL_DOSE_PER_ADMINISTRATION - Actual dose given in a drug administration

OPCS_DELIVERY_CODE - Primary procedure code

ADMINISTRATION_ROUTE - The prescribed route for a drug administration

ADMINISTRATION_DATE - SACT drug administration date

DRUG_GROUP - Drug analysis grouping drug given to patient


MERGED_REGIMEN_ID - Pseudonymised regimen ID

MERGED_CYCLE_ID - Pseudonymised cycle ID

CYCLE_NUMBER - Cycle identifier

MERGED_DATE_OF_CYCLE - Start date of first SACT drug administration in a cycle

OPCS_PROCUREMENT_CODE - Primary procedure (OPCS)

PERF_STATUS_START_OF_CYCLE - Performance status at start of cycle


ENCORE_PATIENT_ID - Pseudonymised patient ID

MERGED_REGIMEN_ID - Pseudonymised regimen ID

HEIGHT_AT_START_OF_REGIMEN - Patient's height (metres [m])

WEIGHT_AT_START_OF_REGIMEN - Patient's weight (kilograms [kg])

INTENT_OF_TREATMENT - Drug treatment intent

DATE_DECISION_TO_TREAT - The date that the consultation between the patient and the care professional took place where a decision was taken to treat a patient with a SACT regimen.

START_DATE_OF_REGIMEN - Start date of the first cycle of a SACT regimen

MAPPED_REGIMEN - Maximally granular mapped regimen

CLINICAL_TRIAL - Clinical trial indicator, identifies if a patient is currently in an active SACT clinical trial

CHEMO_RADIATION - Chemo-radiation indicator

BENCHMARK_GROUP - Regimen grouping (benchmark reports)

LINK_NUMBER - Link number, substitute for NHS number in the real data, which is coded as NHS_NUMBER.


GENEID - Pseudonymised gene ID

TUMOURID - Pseudonymised tumour ID

PATIENTID - Pseudonymised patient ID

GENE_DESC - Gene description

GENE - Gene code

COUNT_TESTS - Count of genetic tests (ids) on this geneid

COUNT_RESULTS - Count of genetic tests results (ids) on this geneid

COUNT_DATE - Number of test dates for this gene

ALL_TESTSTATUSES - Outcome of test is the test status. List of all test statuses ordered by date for this geneid.

OVERALL_TS - Overall test status on whether anything abnormal on this gene. i.e. worst/most conclusive test result As one geneid can have many test statuses, hierarchy assigned in order as listed in notes (teststatus from underlying look up table given in brackets)

NO_OF_AB_GATS - Number of abnormal genetic aberration types, only counting definitive not borderline results.

DNASEQ_GAT - Overall test result for the genetic aberration type 'dna sequence'

METHYL_GAT - Overall test result for the genetic aberration type 'methylation'

EXP_GAT - Overall test result for the genetic aberration type 'expression'

COPYNO_GAT - Overall test result for the genetic aberration type 'copyno'

FUS_TRANS_GAT - Overall test result for the genetic aberration type 'fusion/translocation'

ABNORMAL_GAT - Type of genetic aberration if only one otherwise "multiple" Summary variable for the five different genetic aberration types. If only one of the above five genetic aberration types is abnormal gives it otherwise its 'multiple'.This only counts definitive not borderline results.

NO_OF_SEQ_VARS - Count of sequence variants for this gene

ALL_SEQ_VARS - List of all sequence variants for this gene

SEQ_VAR - Sequence variant for this gene if only one otherwise "multiple"

DATE_OVERALL_TS - Test date for overall test status

BEST_DATE_SOURCE_OVERALL_TS - Source of test (i.e. molecular or pathology feed)

MIN_DATE - First test date for this gene

MAX_DATE - Last test date for this gene

ALL_PRO_IMPS - List of all protein impacts for this gene

NO_PRO_IMPS - Count of protein impacts for this gene

PRO_IMP - Protein impacts for this gene if only one otherwise "multiple"

METHODS - The method of the test.

LAB_NAME - Name of laboratory where test was performed


TUMOURID - Pseudonymised tumour ID

GENDER - Person stated gender

PATIENTID - Pseudonymised patient ID

DIAGNOSISDATETEXT - Diagnosis date

SITE_ICD10_O2_3CHAR - Site of neoplasm (3-character ICD-10/O2 code original version released 2010)

SITE_ICD10_O2 - Site of neoplasm (4-character ICD-10/O2 code original version released 2010)

SITE_ICD10R4_O2_3CHAR_FROM2013 - Site of neoplasm (3-character ICD-10/O2 code 5th revision release 2013) for diagnoses from 2013 onwards

SITE_ICD10R4_O2_FROM2013 - Site of neoplasm (4-character ICD-10/O2 code 5th revision released 2013) for diagnoses from 2013 onwards

SITE_ICD03REV2011 - Site of neoplasm (3-character code ICD-O-3 1st revision released 2013)

SITE_ICD03REV2011_3CHAR - Site of neoplasm (4-character ICD-O-3 1st revision released 2013)

MORPH_ICD10_O2 - Histology of the cancer, in the ICD-10/O2 system

MORPH_ICD03REV2011 - Histology of the cancer, in the ICD-O-3 1st revision released 2013

BEHAVIOUR_ICD10_O2 - Behaviour of the cancer, in the ICD-10/O2 system

BEHAVIOUR_ICD03REV2011 - Behaviour of the cancer, in the ICD-O-3 1st revision released 2013

T_BEST - T stage flagged by the registry as the 'best' T stage

N_BEST - N stage flagged by the registry as the 'best' N stage

M_BEST - M stage flagged by the registry as the 'best' M stage

STAGE_BEST - Best 'registry' stage at diagnosis of the tumour

GRADE - Grade of tumour

AGE - Age at diagnosis

CREG_CODE - Cancer registry catchment area code the patient was resident in when the tumour was diagnosed

STAGE_BEST_SYSTEM - System used to record best registry stage at diagnosis

LATERALITY - Laterality

SCREENINGSTATUSFULL_CODE - Full detailed screening status of the tumour

ER_STATUS - Oestrogen receptor status of the tumour

PR_STATUS - Progesterone receptor status of the tumour

HER2_STATUS - HER2 status of the tumour

QUINTILE_2019 - Measure of deprivation: the population-weighted quintile of income-level deprivation at small area level (LSAD)

DATE_FIRST_SURGERY - Date of first surgical event linked to this tumour recorded in the Cancer Registration treatment table

CANCERCAREPLANINTENT - Intent of treatment as recorded in COSD Cancer Care Plan

PERFORMANCESTATUS - Performance status recorded at diagnosis

CHRL_TOT_27_03 - Charlson comorbidity score

COMORBIDITIES_27_03 - Charlson groups making up the total Charlson score for a lookback of 27 to 3 months.

GLEASON_PRIMARY - Gleason primary pattern

GLEASON_SECONDARY - Gleason secondary pattern

GLEASON_TERTIARY - Gleason tertiary pattern

GLEASON_COMBINED - Combined Gleason primary and secondary scores

# EDA

import pandas as pd
import os

files =  os.listdir("../../data/simulacrum_v2.1.0/Data/")

sheets = pd.ExcelFile("../../data/simulacrum_v2.1.0/Documents/Simulacrum_2_data_dictionaryVer2.1.0_2016-2019.xlsx").sheet_names

len(files)

# dd = pd.read_csv("experiment_data/consolidated_data_dictionary.csv")

# read files
patients = pd.read_csv("../../data/simulacrum_v2.1.0/Data/sim_av_patient.csv", low_memory=False)
prescriptions = pd.read_csv("../../data/simulacrum_v2.1.0/Data/sim_rtds_prescription.csv", low_memory=False)
exposures = pd.read_csv("../../data/simulacrum_v2.1.0/Data/sim_rtds_exposure.csv", low_memory=False)
genes = pd.read_csv("../../data/simulacrum_v2.1.0/Data/sim_av_gene.csv", low_memory=False)
episodes = pd.read_csv("../../data/simulacrum_v2.1.0/Data/sim_rtds_episode.csv", low_memory=False)
outcomes = pd.read_csv("../../data/simulacrum_v2.1.0/Data/sim_sact_outcome.csv", low_memory=False)
cycles = pd.read_csv("../../data/simulacrum_v2.1.0/Data/sim_sact_cycle.csv", low_memory=False)
details = pd.read_csv("../../data/simulacrum_v2.1.0/Data/sim_sact_drug_detail.csv", low_memory=False)
tumours = pd.read_csv("../../data/simulacrum_v2.1.0/Data/sim_av_tumour.csv", low_memory=False)
regimens = pd.read_csv("../../data/simulacrum_v2.1.0/Data/sim_sact_regimen.csv", low_memory=False)
combined = pd.read_csv("../../data/simulacrum_v2.1.0/Data/sim_rtds_combined.csv", low_memory=False)

patients.info()

combined.info()

tumours.info()

#### filter for prostate cancer patients only

prostate_tumours = tumours[tumours['SITE_ICD10R4_O2_FROM2013']== 'C61']
prostate_tumours.to_csv("experiment_data/prostate_tumours.csv", index=False)
prostate_tumours.info()

- There are 179, 478 patients of prostate cancer.

###### Filter all Dataframes for prostate patients only and save results to ``CSV`` to  save on memory loads.

tumours['PATIENTID'].duplicated().any()

prostate_tumours[prostate_tumours['PATIENTID'].duplicated(keep=False)].sort_values('PATIENTID')


merged_df = prostate_tumours.merge(patients, on=("PATIENTID"),  how="left", indicator="merged")

merged_df['merged'].value_counts()

merged_df.info()

merged_df['GENDER_x'].value_counts()

merged_df['GENDER_y'].value_counts()

merged_df["GENDER"] = merged_df["GENDER_x"]

merged_df.drop(['GENDER_x', 'GENDER_y', 'merged'], axis=1, inplace=True)

merged_df.info()

patients[patients["PATIENTID"].isin(prostate_tumours['PATIENTID'])].to_csv("experiment_data/prostate_tumours_patients.csv", index=False)

prescriptions[prescriptions["PATIENTID"].isin(prostate_tumours['PATIENTID'])].to_csv("experiment_data/prostate_tumours_patients_prescriptions.csv", index=False)

exposures[exposures['PATIENTID'].isin(prostate_tumours['PATIENTID'])].to_csv("experiment_data/prostate_tumours_patients_exposures.csv", index=False)

genes[genes['PATIENTID'].isin(prostate_tumours['PATIENTID'])].to_csv("experiment_data/prostate_tumours_patients_genes.csv", index=False)

episodes[episodes['PATIENTID'].isin(prostate_tumours['PATIENTID'])].to_csv("experiment_data/prostate_tumours_patients_episodes.csv", index=False)

exposures[exposures['PATIENTID'].isin(prostate_tumours['PATIENTID'])].to_csv("experiment_data/prostate_tumours_patients_exposures.csv", index=False)

regimens[regimens['ENCORE_PATIENT_ID'].isin(prostate_tumours['PATIENTID'])].to_csv("experiment_data/prostate_tumours_patients_regimens.csv", index=False)

prostate_regimens = regimens[regimens['ENCORE_PATIENT_ID'].isin(prostate_tumours['PATIENTID'])]

prostate_cycles = cycles[cycles["MERGED_REGIMEN_ID"].isin(prostate_regimens["MERGED_REGIMEN_ID"])]

prostate_cycles.to_csv("experiment_data/prostate_tumours_patients_cycles.csv", index=False)

outcomes[outcomes['MERGED_REGIMEN_ID'].isin(prostate_regimens['MERGED_REGIMEN_ID'])].to_csv("experiment_data/prostate_tumours_patients_outcomes.csv", index=False)

details[details["MERGED_CYCLE_ID"].isin(prostate_cycles["MERGED_CYCLE_ID"])].to_csv("experiment_data/prostate_tumours_patients_drugs_details.csv", index=False)
